+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryAutoConfiguration.java
@@ -7,6 +7,7 @@ import io.sentry.IHub;
 import io.sentry.Integration;
 import io.sentry.Sentry;
 import io.sentry.SentryOptions;
+import io.sentry.TransportFactory;
 import io.sentry.protocol.SdkVersion;
 import io.sentry.spring.SentryExceptionResolver;
 import io.sentry.spring.SentryRequestResolver;
@@ -18,7 +19,6 @@ import io.sentry.spring.tracing.SentrySpanAdvice;
 import io.sentry.spring.tracing.SentryTracingFilter;
 import io.sentry.spring.tracing.SentryTransaction;
 import io.sentry.spring.tracing.SentryTransactionAdvice;
-import io.sentry.transport.ITransport;
 import io.sentry.transport.ITransportGate;
 import java.util.List;
 import org.aopalliance.aop.Advice;
@@ -69,7 +69,7 @@ public class SentryAutoConfiguration {
         final @NotNull List<Integration> integrations,
         final @NotNull ObjectProvider<ITransportGate> transportGate,
         final @NotNull List<SentryUserProvider> sentryUserProviders,
-        final @NotNull ObjectProvider<ITransport> transport,
+        final @NotNull ObjectProvider<TransportFactory> transportFactory,
         final @NotNull InAppIncludesResolver inAppPackagesResolver) {
       return options -> {
         beforeSendCallback.ifAvailable(options::setBeforeSend);
@@ -82,7 +82,7 @@ public class SentryAutoConfiguration {
                 options.addEventProcessor(
                     new SentryUserProviderEventProcessor(sentryUserProvider)));
         transportGate.ifAvailable(options::setTransportGate);
-        transport.ifAvailable(options::setTransport);
+        transportFactory.ifAvailable(options::setTransportFactory);
         inAppPackagesResolver.resolveInAppIncludes().forEach(options::addInAppInclude);
       };
     }