+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentrySpringIntegrationTest.kt
@@ -1,20 +1,26 @@
 package io.sentry.spring.boot
 
+import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.anyOrNull
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.reset
 import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.Sentry
+import io.sentry.TransportFactory
 import io.sentry.spring.tracing.SentrySpan
 import io.sentry.test.checkEvent
 import io.sentry.transport.ITransport
 import java.lang.RuntimeException
 import org.assertj.core.api.Assertions.assertThat
 import org.awaitility.kotlin.await
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.boot.web.server.LocalServerPort
 import org.springframework.context.annotation.Bean
@@ -43,11 +49,16 @@ import org.springframework.web.bind.annotation.RestController
 )
 class SentrySpringIntegrationTest {
 
-    @MockBean
+    @Autowired
     lateinit var transport: ITransport
 
     @LocalServerPort
     lateinit var port: Integer
+
+    @Before
+    fun reset() {
+        reset(transport)
+    }
 
     @Test
     fun `attaches request and user information to SentryEvents`() {
@@ -129,7 +140,19 @@ class SentrySpringIntegrationTest {
 }
 
 @SpringBootApplication
-open class App
+open class App {
+    private val transport = mock<ITransport>()
+
+    @Bean
+    open fun mockTransportFactory(): TransportFactory {
+        val factory = mock<TransportFactory>()
+        whenever(factory.create(any())).thenReturn(transport)
+        return factory
+    }
+
+    @Bean
+    open fun mockTransport() = transport
+}
 
 @RestController
 class HelloController(private val helloService: HelloService) {